slimano/nbi/requirements.txt
jsonschema
PyYAML
requests>=2.21
flask>=1.0.2
nameko